backend/api-gateway/pyproject.toml
[project]
name = "orbital-credit-api-gateway"
version = "0.1.0"
description = "Orbital Credit API Gateway"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "psycopg[binary]>=3.1.19,<4.0.0",
  "pystac-client>=0.9.0",
  "planetary-computer>=1.0.0",
  "rasterio>=1.4.4",
]

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0,<9.0.0"
]
